feat: add UserDto, UserDtoMapper, UserController and EwmExceptionHandler classes

+++ ewm-main/src/main/java/ru/practicum/ewmmain/compilation/repository/CompilationRepository.java
@@ -3,5 +3,5 @@ package ru.practicum.ewmmain.compilation.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.ewmmain.compilation.model.Compilation;
 
-public interface CompilationRepository extends JpaRepository<Compilation, Long> {
+public interface CompilationRepository extends JpaRepository<Compilation, Integer> {
 }
