+++ ewm-main/src/main/java/ru/practicum/ewmmain/participationrequest/repository/ParticipationReqRepository.java
@@ -3,5 +3,5 @@ package ru.practicum.ewmmain.participationrequest.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.ewmmain.participationrequest.model.ParticipationRequest;
 
-public interface ParticipationReqRepository extends JpaRepository<ParticipationRequest, Long> {
+public interface ParticipationReqRepository extends JpaRepository<ParticipationRequest, Integer> {
 }
